feat: adds abstract test for controller testing, extracts token duration to properties
