+++ SocialNetwork/src/main/resources/application.properties
@@ -18,4 +18,5 @@ spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 
 social-network.security.secret=secret
+social-network.security.token-duration=600000
 social-network.name=social-network
