+++ SocialNetwork/src/test/java/com/example/SocialNetwork/integration/PostControllerTest.java
@@ -4,55 +4,17 @@ package com.example.SocialNetwork.integration;
 import com.example.SocialNetwork.domain.dto.JwtAuthDTO;
 import com.example.SocialNetwork.domain.dto.PostDTO;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import liquibase.pro.packaged.P;
 import org.junit.Test;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.json.JacksonJsonParser;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
-import static org.junit.Assert.assertEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-@AutoConfigureMockMvc
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class PostControllerTest {
-
-    @Autowired
-    private WebApplicationContext context;
-    @Autowired
-    private MockMvc mvc;
-
-    @BeforeEach
-    public void setup(){
-        this.mvc = MockMvcBuilders
-                .webAppContextSetup(context)
-                .apply(SecurityMockMvcConfigurers.springSecurity())
-                .build();
-    }
-
+public class PostControllerTest extends AbstractControllerTest {
 
     @Test
     public void createPost_noAccessToken_statusIsUnauthorized() throws Exception {
@@ -162,7 +124,7 @@ public class PostControllerTest {
 
         this.mvc.perform(delete("/api/v1/posts/test-post1")
                         .header("Authorization", "Bearer " + accessToken)
-                        )
+                )
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.text", equalTo("test text1")));
     }
